+++ Backend/src/main/java/com/server/crm1/repository/user/UserRepository.java
@@ -14,7 +14,7 @@ public interface UserRepository extends JpaRepository<User, Integer>, UserReposi
 
 	Optional<User> findByEmail(String email);
 
-	List<User> findByOrganizationId(Integer organizationId);
+	List<User> findByOrganizationId(Long organizationId);
 
 	Boolean existsByEmail(String email);
